+++ transaction/_transaction.py
@@ -319,9 +319,9 @@ class Transaction(object):
             finally:
                 del t, v, tb
         else:
-            self._free()
             self._synchronizers.map(lambda s: s.afterCompletion(self))
             self._callAfterCommitHooks(status=True)
+            self._free()
         self.log.debug("commit")
 
     def _saveAndGetCommitishError(self):
@@ -367,12 +367,8 @@ class Transaction(object):
 
     def _callBeforeCommitHooks(self):
         # Call all hooks registered, allowing further registrations
-        # during processing.  Note that calls to addBeforeCommitHook() may
-        # add additional hooks while hooks are running, and iterating over a
-        # growing list is well-defined in Python.
-        for hook, args, kws in self._before_commit:
-            hook(*args, **kws)
-        self._before_commit = []
+        # during processing.
+        self._call_hooks(self._before_commit)
 
     def getAfterCommitHooks(self):
         """ See ITransaction.
@@ -387,34 +383,53 @@ class Transaction(object):
         self._after_commit.append((hook, tuple(args), kws))
 
     def _callAfterCommitHooks(self, status=True):
+        self._call_hooks(self._after_commit,
+                         exc=False, clean=True, prefix_args=(status,))
+
+    def _call_hooks(self, hooks, exc=True, clean=False, prefix_args=()):
+        """call *hooks*.
+
+        If *exc* is true, fail on the first exception; otherwise
+        log the exception and continue.
+
+        If *clean* is true, abort all resources. This is to ensure
+        a clean state should a (after) hook has affected one
+        of the resources.
+
+        *prefix_args* defines additioan arguments prefixed
+        to the arguments provided by the hook definition.
+
+        ``_call_hooks`` supports that a hook adds new hooks.
+        """
         # Avoid to abort anything at the end if no hooks are registred.
-        if not self._after_commit:
+        if not hooks:
             return
-        # Call all hooks registered, allowing further registrations
-        # during processing.  Note that calls to addAterCommitHook() may
-        # add additional hooks while hooks are running, and iterating over a
-        # growing list is well-defined in Python.
-        for hook, args, kws in self._after_commit:
-            # The first argument passed to the hook is a Boolean value,
-            # true if the commit succeeded, or false if the commit aborted.
-            try:
-                hook(status, *args, **kws)
-            except:
-                # We need to catch the exceptions if we want all hooks
-                # to be called
-                self.log.error("Error in after commit hook exec in %s ",
-                               hook, exc_info=sys.exc_info())
-        # The transaction is already committed. It must not have
-        # further effects after the commit.
-        for rm in self._resources:
-            try:
-                rm.abort(self)
-            except:
-                # XXX should we take further actions here ?
-                self.log.error("Error in abort() on manager %s",
-                               rm, exc_info=sys.exc_info())
-        self._after_commit = []
-        self._before_commit = []
+        try:
+            # Call all hooks registered, allowing further registrations
+            # during processing
+            for hook, args, kws in hooks:
+                try:
+                    hook(*(prefix_args + args), **kws)
+                except:
+                    if exc:
+                        raise
+                    # We should not fail
+                    self.log.error("Error in hook exec in %s ",
+                                   hook, exc_info=sys.exc_info())
+        finally:
+            del hooks[:]  # clear hooks
+            if not clean:
+                return
+            # The primary operation has already been performed.
+            # But the hooks execution might have left the resources
+            # in an unclean state. Clean up
+            for rm in self._resources:
+                try:
+                    rm.abort(self)
+                except:
+                    # XXX should we take further actions here ?
+                    self.log.error("Error in abort() on manager %s",
+                                   rm, exc_info=sys.exc_info())
 
     def getBeforeAbortHooks(self):
         """ See ITransaction.
@@ -430,12 +445,8 @@ class Transaction(object):
 
     def _callBeforeAbortHooks(self):
         # Call all hooks registered, allowing further registrations
-        # during processing.  Note that calls to addBeforeAbortHook() may
-        # add additional hooks while hooks are running, and iterating over a
-        # growing list is well-defined in Python.
-        for hook, args, kws in self._before_abort:
-            hook(*args, **kws)
-        self._before_abort = []
+        # during processing.
+        self._call_hooks(self._before_abort, exc=False)
 
     def getAfterAbortHooks(self):
         """ See ITransaction.
@@ -450,32 +461,7 @@ class Transaction(object):
         self._after_abort.append((hook, tuple(args), kws))
 
     def _callAfterAbortHooks(self):
-        # Avoid to abort anything at the end if no hooks are registred.
-        if not self._after_abort:
-            return
-        # Call all hooks registered, allowing further registrations
-        # during processing.  Note that calls to addAterAbortHook() may
-        # add additional hooks while hooks are running, and iterating over a
-        # growing list is well-defined in Python.
-        for hook, args, kws in self._after_abort:
-            try:
-                hook(*args, **kws)
-            except:
-                # We need to catch the exceptions if we want all hooks
-                # to be called
-                self.log.error("Error in after abort hook exec in %s ",
-                               hook, exc_info=sys.exc_info())
-        # The transaction is already abortted. It must not have
-        # further effects after the abort.
-        for rm in self._resources:
-            try:
-                rm.abort(self)
-            except:
-                # XXX should we take further actions here ?
-                self.log.error("Error in abort() on manager %s",
-                               rm, exc_info=sys.exc_info())
-        self._after_abort = []
-        self._before_abort = []
+        self._call_hooks(self._after_abort, clean=True)
 
     def _commitResources(self):
         # Execute the two-phase commit protocol.
@@ -537,6 +523,7 @@ class Transaction(object):
         # to break references---this transaction object will not be returned
         # as the current transaction from its manager after this, and all
         # IDatamanager objects joined to it will forgotten
+        # All hooks are forgotten.
         if self._manager:
             self._manager.free(self)
 
@@ -544,6 +531,11 @@ class Transaction(object):
             delattr(self, '_data')
 
         del self._resources[:]
+
+        del self._before_commit[:]
+        del self._after_commit[:]
+        del self._before_abort[:]
+        del self._after_abort[:]
 
     def data(self, ob):
         try: