+++ CHANGES.rst
@@ -21,6 +21,9 @@
 - Support abort hooks (symmetrically to commit hooks)
   (`#77 <https://github.com/zopefoundation/transaction/issues/77>`_).
 
+- Hooks are now cleared after successfull ``commit`` and ``abort`` to avoid
+  potential cyclic references.
+
 
 2.4.0 (2018-10-23)
 ==================